requirements.txt
aiomysql==0.1.1
fastapi==0.81.0
tortoise-orm==0.19.2
uvicorn==0.18.3
